requirements.txt
Flask~=3.1.1
python-dotenv~=1.1.0
pymongo~=4.13.0
PyDrive~=1.3.1
pdfrw~=0.4
pandas~=2.2.3